requirements.txt
requests
tqdm
beautifulsoup4
PyYAML
